+++ lib/src/dtos/auth_method_provider.g.dart
@@ -9,6 +9,7 @@ part of 'auth_method_provider.dart';
 AuthMethodProvider _$AuthMethodProviderFromJson(Map<String, dynamic> json) =>
     AuthMethodProvider(
       name: json['name'] as String? ?? "",
+      displayName: json['displayName'] as String? ?? "",
       state: json['state'] as String? ?? "",
       codeVerifier: json['codeVerifier'] as String? ?? "",
       codeChallenge: json['codeChallenge'] as String? ?? "",
@@ -19,6 +20,7 @@ AuthMethodProvider _$AuthMethodProviderFromJson(Map<String, dynamic> json) =>
 Map<String, dynamic> _$AuthMethodProviderToJson(AuthMethodProvider instance) =>
     <String, dynamic>{
       'name': instance.name,
+      'displayName': instance.displayName,
       'state': instance.state,
       'codeVerifier': instance.codeVerifier,
       'codeChallenge': instance.codeChallenge,