+++ lib/src/dtos/log_stat.g.dart
@@ -1,19 +1,17 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'log_request_stat.dart';
+part of 'log_stat.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-LogRequestStat _$LogRequestStatFromJson(Map<String, dynamic> json) =>
-    LogRequestStat(
+LogStat _$LogStatFromJson(Map<String, dynamic> json) => LogStat(
       total: json['total'] as int? ?? 0,
       date: json['date'] as String? ?? "",
     );
 
-Map<String, dynamic> _$LogRequestStatToJson(LogRequestStat instance) =>
-    <String, dynamic>{
+Map<String, dynamic> _$LogStatToJson(LogStat instance) => <String, dynamic>{
       'total': instance.total,
       'date': instance.date,
     };