+++ test/dtos/auth_methods_list_test.dart
@@ -10,6 +10,7 @@ void main() {
         "authProviders": [
           {
             "name": "test_name",
+            "displayName": "test_displayName",
             "state": "test_state",
             "codeVerifier": "test_codeVerifier",
             "codeChallenge": "test_codeChallenge",